+++ lib/appsflyer_sdk.dart
@@ -31,3 +31,4 @@ part 'src/purchase_connector/models/ios_error.dart';
 part 'src/appsflyer_consent.dart';
 part 'src/appsflyer_request_listener.dart';
 part 'appsflyer_sdk.g.dart';
+part 'src/appsflyer_ad_revenue_data.dart';