+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Versions
+## 6.15.2
+- Fixed NullPointerException issue on Android that some clients had.
+- Fixed Android MediationNetwork enum issue.
+- Update iOS version to 6.15.3
+- Update Android version to 6.15.2
+## 6.15.1
+- Implementation of the new logAdRevenue API for iOS and Android
+- Documentation update for the new logAdRevenue API
+- Update iOS version to 6.15.1
+- Update Android version to 6.15.1
 ## 6.14.3
 - Fixed mapOptions issue with manualStart
 - Inherit Privacy Manifest from the native iOS SDK via Cocoapods